TODO: dsdb: Do not refresh the schema using the wrong event context
What we now do is have the refresh function and module be on a
seperate object to the schema, only referring to the data and
not excuting on the original ldb and event loop.

That is, we never use another ldb context when calling the
refresh function, by binding the refresh handler to the
ldb and not the schema.

Andrew Bartlett

Change-Id: I5c323dda743cf5858badd01147fda6227599bc16
Signed-off-by: Andrew Bartlett <abartlet@samba.org>

TODO:

Please use:

if (mem_ctx == NULL)

and

if (schema != NULL)

==============

Please wrap debug statements so they do not exceed 80 chars.

==============

In dsdb_schema_from_db() you remove

ret = dsdb_set_schema(ldb, (*schema));

but check for the return code is still there ...

==============

In schema_load_init() I would prefer if you introduce helper variables.

void *o;

o = ldb_get_opaque(ldb, "readOnlySchema");
if (o != NULL) {
    ...
}

Same for dsdb_get_schema() calls in here ...

==============

In dsdb_set_global_schema()

if (global_schema == NULL)

==============

In dsdb_get_schema()

if (p != NULL)

==============