TODO: dsdb: Do not store a struct ldb_dn in struct schema_data
The issue is that the DN contains a pointer to the ldb it belongs to,
and if this is not kept around long enough, we might reference memory
after it is de-allocated.

Andrew Bartlett

Change-Id: I040a6c37a3164b3309f370e32e598dd56b1a1bbb
Signed-off-by: Andrew Bartlett <abartlet@samba.org>

TODO:
As you're already changing the code, we should not do a check inside of
the if-clause.

Please use or introduce variables. It makes it easier to debug the code.

cmp = ldb_dn_compare(partition->dn, ldb_get_schema_basedn(service->samdb));
if (cmp != 0) {
    error
}