util/debug: debug_list_class_names_and_levels - accept explicite memory context
This implementation also adds a bit change of behavior for
debug_list_class_names_and_levels() as previously, it would
return NULL when debug_num_classes == 0. NULL answer is treated
as "OOM" by caller - eg. debuglevel_message().

With this patch though, when debug_num_classes == 0 we are going
to return just an empty string. Which, imho should be correct
behavior.

Signed-off-by: Kamen Mazdrashki <kamenim@samba.org>